Remove widgetmaps.h
