+++ include/reactive/debug/drawkeyboardinputs.h
@@ -1,14 +1,14 @@
 #pragma once
 
-#include "reactive/widget/widgettransformer.h"
+#include "reactive/widget/ondraw.h"
 #include "reactive/widget/bindkeyboardinputs.h"
 #include "reactive/widget/binddrawcontext.h"
+#include "reactive/widget/widgettransformer.h"
 
+#include "reactive/shapes.h"
 #include "reactive/widgetfactory.h"
 
 #include "reactive/signal/map.h"
-
-#include "reactive/shapes.h"
 
 #include <avg/pathbuilder.h>
