+++ include/reactive/dynamicbox.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "widget/addwidgets.h"
 #include "widget/bindsize.h"
 #include "widget/binddrawcontext.h"
 #include "widget/widgetobject.h"
