+++ src/reactive/widget/widgetobject.cpp
@@ -1,5 +1,5 @@
 #include "widget/widgetobject.h"
-#include "widgetmaps.h"
+#include "widget/transform.h"
 
 #include <pmr/new_delete_resource.h>
